requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0
six>=1.9.0
oslo.log
oslo.config
oslo.utils
oslo.service
oslo.db
taskflow
keystoneauth1
python-novaclient
stomp.py>=4.1.13
SQLAlchemy-Utils
